switch

+++ home/programs/scripts/sr.sh
@@ -27,20 +27,55 @@ do
 
 		elif [ $(( currentRate % 48000 )) -eq 0 ];
 		then
-			pid=$(pgrep qpwgraph)
-			kill "$pid"
-			pw-metadata -n settings 0 clock.force-rate 48000
-			pwrate=${currentRate}
+			if [[ " ${allowedRates[*]}" =~ [[:space:]]192000[[:space:]] ]];
+			then
+				pid=$(pgrep qpwgraph)
+				kill "$pid"
+				pw-metadata -n settings 0 clock.force-rate 192000
+				pwrate=${currentRate}
+				hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+			elif [[ " ${allowedRates[*]}" =~ [[:space:]]96000[[:space:]] ]];
+			then
+				pid=$(pgrep qpwgraph)
+				kill "$pid"
+				pw-metadata -n settings 0 clock.force-rate 96000
+				pwrate=${currentRate}
 
-			hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+				hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+			elif [[ " ${allowedRates[*]}" =~ [[:space:]]48000[[:space:]] ]];
+			then
+				pid=$(pgrep qpwgraph)
+				kill "$pid"
+				pw-metadata -n settings 0 clock.force-rate 48000
+				pwrate=${currentRate}
+
+				hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+			else
+				pid=$(pgrep qpwgraph)
+				kill "$pid"
+				pw-metadata -n settings 0 clock.force-rate 44100
+				pwrate=${currentRate}
+
+				hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+			fi
 
 		else
-			pid=$(pgrep qpwgraph)
-			kill "$pid"
-			pw-metadata -n settings 0 clock.force-rate 41000
-			pwrate=${currentRate}
+			if [[ " ${allowedRates[*]}" =~ [[:space:]]88200[[:space:]] ]];
+			then
+				pid=$(pgrep qpwgraph)
+				kill "$pid"
+				pw-metadata -n settings 0 clock.force-rate 88200
+				pwrate=${currentRate}
 
-			hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+				hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+			else
+				pid=$(pgrep qpwgraph)
+				kill "$pid"
+				pw-metadata -n settings 0 clock.force-rate 44100
+				pwrate=${currentRate}
+
+				hyprctl dispatch exec "[workspace 4 silent]" uwsm app -- qpwgraph
+			fi
 		fi
 
 	fi
